+++ core/src/main/java/react4j/core/ReactConfig.java
@@ -43,7 +43,7 @@ public final class ReactConfig
 
   /**
    * Return true if we should check that the user interacts with React component in a way compatible with the state.
-   * i.e. setInitialState should only be invoked at the start, setState should not be invoked while componentWillUpdate
+   * i.e. setInitialState should only be invoked at the start, setState should not be invoked while render
    * is running etc.
    *
    * @return true to enable invariant checking about how we interact with native react component.