+++ processor/src/main/java/react4j/processor/ReactProcessor.java
@@ -52,7 +52,6 @@ public final class ReactProcessor
                    "componentDidUpdate",
                    "componentWillReceiveProps",
                    "componentWillUnmount",
-                   "componentWillUpdate",
                    "componentDidCatch",
                    "shouldComponentUpdate" );